manual update complete
the manual entry and calculation has been updated. the user may enter a
value based on which lottery is currently chosen and will receive the
calculated annual payments based on that amount. if they wish to change
lottery selection, the value will default to the current lottery value.

Now to add a little bit of style...